requirements.txt
Django==4.2.3
mysqlclient==2.1.1
dj-database-url==1.2.0
djangorestframework==3.14.0
gunicorn==20.1.0
django-cors-headers==3.14.0
django-ajax-selects==2.2.0
whitenoise==6.5.0
